implemented auth controller
